tdf#134188 tdf#114223 Don't override vertical text button visibility
i.e. restore the fix for tdf#33356, which was removed in commit
eeb2d19e77d6dc47c68e8ba0920a02cf64a1247b ("Resolves: tdf#130159
hide writedirection unless CTL is enabled"). But do it in a way
that doesn't introduce tdf#114223 (a regression of tdf#33356).

We can assume that the visibility state of the button at the
point of control initialization is the one set with toolbar
customization, as any customization triggers clearing of the
whole toolbar, and reconstructing from scratch.

Change-Id: I2cdf18e144dd6eb99b8c1e1da968beacddbb3728
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/98317
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>