SVG export: fix lost semi-transparent text on shapes
Extend SVGTextWriter::setTextPosition(), so when it looks for a text
action in a metafile, it recurses into transparency groups, so the text
is not lost.

Extract part of SVGActionWriter::ImplWriteMask() into a new StartMask(),
so we can detect the case when the transparency group has a constant
alpha, i.e. no complex mask is needed, just an opacity value.

When looking for text, remember if we saw a request for text opacity and
make the transparency group writing in SVGActionWriter::ImplWriteMask()
conditional to avoid duplication. This is needed because once we're
inside <text>, we don't want to write an invalid transparency group via
<g>, rather we want a fill-opacity on the existing <tspan>.

With this, the SVG export is on par with PDF export for semi-transparent
shape text.

Change-Id: If43b0ab3446015299acc4b37590358867c5fac5f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/98937
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins