tdf#68326 writerfilter: default style type is paragraph
Don't ignore styles that don't specify their type.
They are paragraph styles.

17.7.4.17 style (Style Definition) w:type
"If this attribute is not specified, then the
default value shall be assumed to be paragraph."

No existing unit tests are missing this,
and the bug example is a horrible document,
so I'm going without a unit test for this one.

Change-Id: Iaf97dc2d0a5701915f08b7be91fb983cf941380c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/98980
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>