tdf#130559: don't export preview view data to ODS
... because trying to activate this view on load will throw and crash

Unlike in Writer (SwPagePreview), Calc's ScPreviewShell overrides
WriteUserDataSequence to export the view data. This was implemented
in commit 1f51730f3c5b05c3afaa35431c925482e8152783. That does not
give any clue why is this necessary, so I don't want to revert it.
So to pass the flag to XViewDataSupplier::getViewData, this patch
sets the flag in current context using comphelper::NewFlagContext.
This allows to only skip providing the data conditionally.

Change-Id: I95d70b30b0b8762d2b5c703c3e572064d0b2ef75
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/99441
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>