tdf#134959 move focus on activate, but leave focus on value-changed
if the user presses enter to activate the Entry try to go to the page, and on
success we move focus to the document.

if the SpinButton changed value a Timer was launched, when it times-out, move
to the desired page, but leave focus where it currently is, i.e. typically
remaining in the spinbutton, or whatever other widget the user moved to in the
meantime

This replaces always grabbing focus to the document on a page-change and
re-grabbing focus back to the spinbutton.

Change-Id: I597597aedb1753aee755671a2a55779ca6196c09
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/99611
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>