tdf#135216 sw MSexport: save LeftPageDesc as left page style
GetMaster() is for right pages. I assume this was a
copy/paste mistake which came in already at initial import.

And since left/right are alternating, they always use
the first version of the header/footer. Although the UI
doesn't allow changing these, they still COULD be different,
as is the case with the unit test.

No existing unit tests even touch this code,
so not a common situation at all.

Change-Id: I9e3720ddf34a8f7e08ce196780fbe16e8c88940b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/99628
Tested-by: Justin Luth <justin_luth@sil.org>
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-by: Justin Luth <justin_luth@sil.org>