vcl: save and load the page number when swapping the graphic
PDF vector graphic includes a page number, of the page that the
graphic is rendering. This however isn't remembered when swapping
out and back in the graphic, because the serialization format
doesn't include it.

This adds a version 2 of the serialization format, with an
additional page number (page index) attribute.

Also changes the GraphicTest to account for an additional 4 bytes
written and the change of the checksum.

Change-Id: Ic0fbfc4ad983f7880e06956da3b4664bd4b610d4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/100836
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>