oox smartart: consider rules when scaling in linear layout
The bugdoc has an arrow shape which is 100% wide, and there are multiple
shapes before it, which also have a 100% wide constraint. The reason
PowerPoint scales down the shapes (but not the arrow) is because rules
declare it should happen this way.

So start taking rules into account in linear layouts.

Change-Id: I352443277e88be0eb711659489587127727a258f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/99855
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins