NFC sw MS export: flatten useless aTmpFlyIter loop
The loop did nothing, since aTmpFlyIter was never used.
IsAnchorLinkedToThisNode() is only called in one place
to test whether to bPostponeWritingText.

One option would have been to actually use aTmpFlyIter
instead of maFlyIter, which would seem to make the most sense,
and was likely @synerzip's original intention.
(And in that case, it should return false as soon as
nNodePos < nAnchorPos since this is a sorted list.)

However, this whole concept of PostponeWritingText is ripe
for regressions etc. I don't want to introduce an even
larger number of postponed cases, so instead I chose to do
option two, which is to just re-write the function in the
way that it actually operated. So this is supposed to
be in a No-Functional-Change cleanup.

Change-Id: I227e2ecd2aca82b03b54ae3924db91933ca2c930
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/100374
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>