related tdf#135343 writerfilter: document broken nextColumn
Fortunately, it is rather hard in Word to insert a nextColumn
section-break, so this situation is rarely seen,
and bug 135343's fix made it even more irrelevant.
The fix includes unit tests demonstrating this problem.

The column break should be inserted around m_xStartingRange,
but the section should start from pPrevSection->m_xStartingRange
and be extended include this section. But I have no idea
how to do that.

At least documenting that something is completely broken
should be very useful for anyone drawn to this code.

Change-Id: I878f99601bb2181500d215a0440a34a62f3840f8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/100461
Tested-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>