tdf#77794 writerfilter: compat15 - always bLayoutInCell
Based on observation (no documentation found),
MS Word 2013 ignores the actual value of LayoutInCell
and always considers it to be true. The UI shows it,
but it is grayed out when CompatibilityMode == 15.

The code's default value of bLayoutInCell is also true,
so compatMode only needs to be checked if the property
is actually set in the document.

I probably could have left off the IsInTable() check,
but decided to just be as safe as possible.
I assume it doesn't apply, but LayoutInCell sets
a IsFollowingTextFlow property - which might have
meaning outside of a table in SW - I don't know.

Change-Id: Ief073ea96c6c61efb3bcbbc11d114d739b04c3af
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/100651
Tested-by: Jenkins
Reviewed-by: László Németh <nemeth@numbertext.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>