Make "Location" line stick to the bottom of the table of JREs
At least on Linux with SAL_USE_VCLPLUGIN=gtk3, this already happened to be laid
out well, but e.g. on Linux with SAL_USE_VCLPLUGIN=gen and on macOS, there was
blank space between the table and the "Location" line, and the latter was rather
kept close to the following "Optional Features" section.

(There still appears to be variance among the various backends how the two "Java
Options" and "Optional Features" sections are spread out vertically when the
dialog is resized.)

Change-Id: I58ba9fc0e29916f0fa0efb7c3180da058ba72f0b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/100921
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>