Related: tdf#107548 Make sure embedded images have higher priority
As documented in the schema, the dev guide and a comment in
AddonsOptions_Impl::ReadImageData. Noticed the problem while testing
the ProtocolHandlerAddon sdk example which has the ImageSmall props
defined yet didn't show the icons. Turned out this was a result of
the previous commit that fixed ImageSmallURL, which started now to
override ImageSmall.

Change-Id: I0a9eb6b13b73a60efc801905601894c862d68cba
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/101666
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>