intermittent forecast.ets.add.fods failure
valgrinding the forecast.ets.add.fods failure gives...

a) mnSmplInPrd is 9
b) ScETSForecastCalculation::prefillPerIdx:476 fills mpPerIdx to index 8 with values.
c) ScETSForecastCalculation::refill:791 (the bAdditive branch)
   due to nIdx = ( i > mnSmplInPrd ? i - mnSmplInPrd : i ) with i of 9 gives nIdx of 9 and
   uninit mpPerIdx[9] is used in the calculation.
d) At line 799 (the !bAdditive branch)
   due to nIdx = ( i >= mnSmplInPrd ? i - mnSmplInPrd : i ) the nIdx used would be 0 not 9
   under the same circumstances

extending the initialization of prefillPerIdx to zeroing out the next index gives the same
results as we usually get and test passes without a need to change it while avoiding the use
of an uninitialized value.

Change-Id: Ib4629b28ecf2389c0c7611137b756d4d2d9fd700
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/101997
Tested-by: Caolán McNamara <caolanm@redhat.com>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>