tdf#136587 writerfilter: prevent exception in applyStyleSheets
The only thing that a style needs to be valid is a name.
In this case, there was no name imported, so the
exception killed the import of the rest of the styles.

As a test, I forced the empty name to "junk"
and that also worked, so I am fairly confident
that this is all that is necessary to prevent
an exception.

This improves on 7.1 commit for tdf#68326.

NOTE: this unit test does prove that LO is not importing
the Normal style at all. It should import with
fontsize 32. So a proper reading of the RTF file
would also have solved the bug.

Change-Id: I9258bf985a501952dcd5aa52e54f920c99e509f9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/102315
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>