abort if Skia code detects problems with Vulkan
The document from tdf#136244 in low memory conditions (32bit Windows)
may be sometimes rendered without the alpha channel, i.e. Skia/Vulkan
apparently render just the data bitmap but not the alpha bitmap.
So check the status of the GPU context and abort if there's a problem.
I've considered trying to recover by switching to Raster, but
e.g. in this specific case something has already been drawn and it
can't be done. Another problem in this specific case is that the whole
app is running low on memory and copying the data from the GPU would
fail anyway. So just fail hard to avoid silent data loss.

Change-Id: If05f9c65b160ad5e7b9407b9f809606a1392d740
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/103206
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>